Keep all exit statements in a single place
